requirements.txt
openai-agents==0.0.12
dataclass_csv
python-dotenv
pydantic
jinja2
pyyaml